DFM Forecasts

This function produces h-step ahead forecasts of both the factors and the data, with an option to also forecast autocorrelated residuals with a univariate method and produce a combined forecast.

# S3 method for dfm
predict(
  object,
  h = 10L,
  method = switch(object$em.method, none = "2s", "qml"),
  standardized = TRUE,
  resFUN = NULL,
  resAC = 0.1,
  ...
)

# S3 method for dfm_forecast
print(x, digits = 4L, ...)

# S3 method for dfm_forecast
plot(
  x,
  main = paste(x$h, "Period Ahead DFM Forecast"),
  xlab = "Time",
  ylab = "Standardized Data",
  factors = seq_len(ncol(x$F)),
  scale.factors = TRUE,
  factor.col = rainbow(length(factors)),
  factor.lwd = 1.5,
  fcst.lty = "dashed",
  data.col = c("grey85", "grey65"),
  legend = TRUE,
  legend.items = paste0("f", factors),
  grid = FALSE,
  vline = TRUE,
  vline.lty = "dotted",
  vline.col = "black",
  ...
)

# S3 method for dfm_forecast
as.data.frame(
  x,
  ...,
  use = c("factors", "data", "both"),
  pivot = c("long", "wide"),
  time = seq_len(nrow(x$F) + x$h),
  stringsAsFactors = TRUE
)

Arguments

object

an object of class 'dfm'.

h

integer. The forecast horizon.

method

character. The factor estimates to use: one of "qml", "2s" or "pca".

standardized

logical. FALSE will return data forecasts on the original scale.

resFUN

an (optional) function to compute a univariate forecast of the residuals. The function needs to have a second argument providing the forecast horizon (h) and return a vector or forecasts. See Examples.

resAC

numeric. Threshold for residual autocorrelation to apply resFUN: only residual series where AC1 > resAC will be forecasted.

...

not used.

x

an object class 'dfm_forecast'.

digits

integer. The number of digits to print out.

main, xlab, ylab

character. Graphical parameters passed to ts.plot.

factors

integers indicating which factors to display. Setting this to NA, NULL or 0 will omit factor plots.

scale.factors

logical. Standardize factor estimates, this usually improves the plot since the factor estimates corresponding to the greatest PCA eigenvalues tend to have a greater variance than the data.

factor.col, factor.lwd

graphical parameters affecting the colour and line width of factor estimates plots. See par.

fcst.lty

integer or character giving the line type of the forecasts of factors and data. See par.

data.col

character vector of length 2 indicating the colours of historical data and forecasts of that data. Setting this to NA, NULL or "" will not plot data and data forecasts.

legend

logical. TRUE draws a legend in the top-left of the chart.

legend.items

character names of factors for the legend.

grid

logical. TRUE draws a grid on the background of the plot.

vline

logical. TRUE draws a vertical line deliminating historical data and forecasts.

vline.lty, vline.col

graphical parameters affecting the appearance of the vertical line. See par.

use

character. Which forecasts to use "factors", "data" or "both".

pivot

character. The orientation of the frame: "long" or "wide".

time

a vector identifying the time dimension, must be of length T + h, or NULL to omit a time variable.

stringsAsFactors

logical. If TRUE and pivot = "long" the 'Variable' column is created as a factor. Same as option to as.data.frame.table.

Examples

dfm <- DFM(diff(Seatbelts[, 1:7], lag = 12), 3, 3)
#> Converged after 26 iterations.
predict(dfm)
#> 10 Step Ahead Forecast from Dynamic Factor Model
#> 
#> Factor Forecasts
#>        f1     f2      f3
#> 1  1.9460 0.3965  0.3717
#> 2  1.5853 0.0332  0.1615
#> 3  1.1007 0.0733 -0.0048
#> 4  0.8365 0.0753 -0.0221
#> 5  0.5711 0.0577 -0.0592
#> 6  0.4423 0.0499 -0.0304
#> 7  0.3203 0.0456 -0.0134
#> 8  0.2491 0.0375  0.0112
#> 9  0.1874 0.0260  0.0258
#> 10 0.1429 0.0187  0.0368
#> 
#> Series Forecasts
#>    DriversKilled drivers  front   rear    kms PetrolPrice VanKilled
#> 1         0.8302  1.0732 1.0180 0.7032 0.3802     -0.3165    0.3913
#> 2         0.6488  0.8477 0.8167 0.5682 0.3034     -0.3033    0.0059
#> 3         0.4291  0.5622 0.5452 0.3734 0.2209     -0.2953    0.0235
#> 4         0.3228  0.4229 0.4104 0.2798 0.1702     -0.2396    0.0322
#> 5         0.2122  0.2790 0.2725 0.1838 0.1194     -0.1935    0.0171
#> 6         0.1672  0.2194 0.2134 0.1444 0.0917     -0.1406    0.0221
#> 7         0.1228  0.1606 0.1557 0.1054 0.0664     -0.0977    0.0272
#> 8         0.0995  0.1296 0.1247 0.0852 0.0503     -0.0622    0.0285
#> 9         0.0780  0.1013 0.0968 0.0669 0.0366     -0.0350    0.0235
#> 10        0.0627  0.0811 0.0768 0.0537 0.0267     -0.0153    0.0210
fcfun <- function(x, h) predict(ar(x), n.ahead = h)$pred
predict(dfm, resFUN = fcfun)
#> 10 Step Ahead Forecast from Dynamic Factor Model
#> 
#> Factor Forecasts
#>        f1     f2      f3
#> 1  1.9460 0.3965  0.3717
#> 2  1.5853 0.0332  0.1615
#> 3  1.1007 0.0733 -0.0048
#> 4  0.8365 0.0753 -0.0221
#> 5  0.5711 0.0577 -0.0592
#> 6  0.4423 0.0499 -0.0304
#> 7  0.3203 0.0456 -0.0134
#> 8  0.2491 0.0375  0.0112
#> 9  0.1874 0.0260  0.0258
#> 10 0.1429 0.0187  0.0368
#> 
#> Series Forecasts
#>    DriversKilled drivers  front    rear     kms PetrolPrice VanKilled
#> 1         0.8302  1.0732 1.0180  0.7391  0.7658     -0.3181    0.3913
#> 2         0.6488  0.8477 0.8167  0.8697  0.5170     -0.2851    0.0059
#> 3         0.4291  0.5622 0.5452  0.3129  0.7226     -0.3393    0.0235
#> 4         0.3228  0.4229 0.4104  0.1316 -0.3464     -0.2034    0.0322
#> 5         0.2122  0.2790 0.2725  0.4248  0.2259     -0.2155    0.0171
#> 6         0.1672  0.2194 0.2134 -0.3303  0.1182     -0.1349    0.0221
#> 7         0.1228  0.1606 0.1557  0.3344  0.2050     -0.0916    0.0272
#> 8         0.0995  0.1296 0.1247 -0.3911  0.1850     -0.0876    0.0285
#> 9         0.0780  0.1013 0.0968  0.5414 -0.3656     -0.0048    0.0235
#> 10        0.0627  0.0811 0.0768  0.7651  0.0926     -0.0260    0.0210